refactor proxy/mappers to middleware
